Fixed Hibernate types for Postgres specific types.
The json and jsonb types now have a user type. This user type wraps provided values to/from a PGobject with the correct type. The exposed type is still String. The reason for this is that this way we don't have to serialize/deserialize the JSON unnecessarily. The gmtdata-postgres library needs to be added to use these types.

A fix has also been added for the UUID type. The default Hibernate UUID type does not work with Postgres and a specific Type annotation needs to be added. This type annotation is now added at runtime when the DbConnection is created.